# ui/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scanopy API",
    "description": "Network topology discovery and visualization API",
    "license": {
      "name": "Dual (AGPL3.0, Commercial License Available)"
    },
    "version": "0.12.5"
  },
  "paths": {
...
    "/api/daemons/{id}": {
      "get": {
        "tags": [
          "daemons"
        ],
        "summary": "Get daemon by ID",
        "operationId": "get_daemon_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Daemon ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daemon found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Daemon"
                }
              }
            }
          },
          "404": {
            "description": "Daemon not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "put": {
        "tags": [
          "daemons"
        ],
        "summary": "Update daemon",
        "operationId": "update_daemon",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Daemon ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Daemon"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Daemon updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Daemon"
                }
              }
            }
          },
          "404": {
            "description": "Daemon not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "daemons"
        ],
        "summary": "Delete daemon",
        "operationId": "delete_daemon",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Daemon ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daemon deleted"
          },
          "404": {
            "description": "Daemon not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/discovery": {
      "get": {
        "tags": [
          "discoveries"
        ],
        "summary": "List all discoveries",
        "operationId": "list_discoveries",
        "parameters": [
          {
            "name": "network_id",
            "in": "query",
            "description": "Filter by network ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          },
          {
            "name": "daemon_id",
            "in": "query",
            "description": "Filter by daemon ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of discoveries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Discovery"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "discoveries"
        ],
        "summary": "Create a new discovery",
        "operationId": "create_discovery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Discovery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Discovery created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Discovery"
                }
              }
            }
          },
          "400": {
            "description": "Invalid subnet network"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/discovery/start-session": {
      "post": {
        "tags": [
          "discoveries"
        ],
        "summary": "Start a discovery session",
        "operationId": "start_session",
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Discovery session started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryUpdatePayload"
                }
              }
            }
          },
          "404": {
            "description": "Discovery not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Daemon": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DaemonBase"
          },
          {
            "type": "object",
            "properties": {
              "created_at": {
                "type": "string",
                "format": "date-time",
                "readOnly": true
              },
              "id": {
                "type": "string",
                "format": "uuid",
                "readOnly": true
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "readOnly": true
              }
            }
          }
        ]
      },
      "DaemonBase": {
        "type": "object",
        "required": [
          "host_id",
          "network_id",
          "url",
          "mode",
          "name"
        ],
        "properties": {
          "capabilities": {
            "$ref": "#/components/schemas/DaemonCapabilities"
          },
          "host_id": {
            "type": "string",
            "format": "uuid"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "mode": {
            "$ref": "#/components/schemas/DaemonMode"
          },
          "name": {
            "type": "string"
          },
          "network_id": {
            "type": "string",
            "format": "uuid"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "url": {
            "type": "string",
            "readOnly": true
          }
        }
      },
      "DaemonCapabilities": {
        "type": "object",
        "description": "Daemon capabilities",
        "properties": {
          "has_docker_socket": {
            "type": "boolean"
          },
          "interfaced_subnet_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "DaemonMode": {
        "type": "string",
        "enum": [
          "Push",
          "Pull"
        ]
      },
      "Discovery": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DiscoveryBase"
          },
          {
            "type": "object",
            "properties": {
              "created_at": {
                "type": "string",
                "format": "date-time",
                "readOnly": true
              },
              "id": {
                "type": "string",
                "format": "uuid",
                "readOnly": true
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "readOnly": true
              }
            }
          }
        ]
      },
      "DiscoveryBase": {
        "type": "object",
        "required": [
          "discovery_type",
          "run_type",
          "name",
          "daemon_id",
          "network_id"
        ],
        "properties": {
          "daemon_id": {
            "type": "string",
            "format": "uuid"
          },
          "discovery_type": {
            "$ref": "#/components/schemas/DiscoveryType"
          },
          "name": {
            "type": "string"
          },
          "network_id": {
            "type": "string",
            "format": "uuid"
          },
          "run_type": {
            "$ref": "#/components/schemas/RunType"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "DiscoveryPhase": {
        "type": "string",
        "enum": [
          "Pending",
          "Starting",
          "Started",
          "Scanning",
          "Complete",
          "Failed",
          "Cancelled"
        ]
      },
      "DiscoveryType": {
        "oneOf": [
          {
            "type": "object",
            "title": "SelfReport",
            "required": [
              "host_id",
              "type"
            ],
            "properties": {
              "host_id": {
                "type": "string",
                "format": "uuid"
              },
              "type": {
                "type": "string",
                "enum": [
                  "SelfReport"
                ]
              }
            }
          },
          {
            "type": "object",
            "title": "Network",
            "required": [
              "type"
            ],
            "properties": {
              "host_naming_fallback": {
                "$ref": "#/components/schemas/HostNamingFallback"
              },
              "subnet_ids": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "type": {
                "type": "string",
                "enum": [
                  "Network"
                ]
              }
            }
          },
          {
            "type": "object",
            "title": "Docker",
            "required": [
              "host_id",
              "type"
            ],
            "properties": {
              "host_id": {
                "type": "string",
                "format": "uuid"
              },
              "host_naming_fallback": {
                "$ref": "#/components/schemas/HostNamingFallback"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Docker"
                ]
              }
            }
          }
        ]
      },
      "DiscoveryUpdatePayload": {
        "type": "object",
        "description": "Progress update from daemon to server during discovery",
        "required": [
          "session_id",
          "daemon_id",
          "network_id",
          "phase",
          "discovery_type",
          "progress"
        ],
        "properties": {
          "daemon_id": {
            "type": "string",
            "format": "uuid"
          },
          "discovery_type": {
            "$ref": "#/components/schemas/DiscoveryType"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "finished_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "network_id": {
            "type": "string",
            "format": "uuid"
          },
          "phase": {
            "$ref": "#/components/schemas/DiscoveryPhase"
          },
          "progress": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "session_id": {
            "type": "string",
            "format": "uuid"
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "HostNamingFallback": {
        "type": "string",
        "enum": [
          "Ip",
          "BestService"
        ]
      },
...
      "RunType": {
        "oneOf": [
          {
            "type": "object",
            "title": "Scheduled",
            "required": [
              "cron_schedule",
              "enabled",
              "type"
            ],
            "properties": {
              "cron_schedule": {
                "type": "string"
              },
              "enabled": {
                "type": "boolean"
              },
              "last_run": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Scheduled"
                ]
              }
            }
          },
          {
            "type": "object",
            "title": "Historical",
            "required": [
              "results",
              "type"
            ],
            "properties": {
              "results": {
                "$ref": "#/components/schemas/DiscoveryUpdatePayload"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Historical"
                ]
              }
            }
          },
          {
            "type": "object",
            "title": "AdHoc",
            "required": [
              "type"
            ],
            "properties": {
              "last_run": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              },
              "type": {
                "type": "string",
                "enum": [
                  "AdHoc"
                ]
              }
            }
          }
        ],
        "description": "Note: DiscoveryUpdatePayload field is skipped from TS export as it's only used internally"
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for daemon authentication. Generate keys in the web UI under Settings > API Keys."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  }
}